Add some validation Implement Bearer token authentication Implement Integration API tests for blog controller

+++ src/main/java/org/fsq/blogapi/controller/BlogController.java
@@ -1,4 +1,5 @@
 package org.fsq.blogapi.controller;
+import jakarta.validation.Valid;
 import org.bson.types.ObjectId;
 import org.fsq.blogapi.model.Blog;
 import org.fsq.blogapi.model.Response;
@@ -8,52 +9,66 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
 
 @RestController
-@RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE)
+@RequestMapping(value = "/blogs",produces = MediaType.APPLICATION_JSON_VALUE)
 public class BlogController {
 
     private final BlogRepository blogRepository;
-
-    @Value("${spring.profiles.active}")
-    private String profile;
 
     @Autowired
     public BlogController(BlogRepository blogRepository) {
         this.blogRepository = blogRepository;
     }
 
-    @GetMapping("/blogs")
+    @GetMapping
     public List<Blog> getBlogs() {
-        return blogRepository.findAll();
+        List<Blog> blogs = blogRepository.findAll();
+        blogs.forEach(blog -> {
+            if (blog.getUser() != null) {
+                blog.getUser().setBlogs(null);
+            }
+        });
+        return blogs;
     }
 
-    @GetMapping("/blogs/{id}")
+    @GetMapping("/{id}")
     public ResponseEntity<Blog> getBlog(@PathVariable ObjectId id) {
         Optional<Blog> result = blogRepository.findById(id);
         return result.map(blog -> ResponseEntity.status(HttpStatus.OK).body(blog)).orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).build());
     }
 
-    @PostMapping("/blogs")
-    public ResponseEntity<Response> addBlog(@RequestBody Blog blog) {
-        System.out.println(profile);
+    @PostMapping
+    public ResponseEntity<Response> addBlog(@Valid @RequestBody Blog blog) {
+        if (blog.getLikes() == 0) {
+            blog.setLikes(0);
+        }
+
         blogRepository.save(blog);
         Response response = new Response(HttpStatus.CREATED,"Blog successfully added.");
         return ResponseEntity.status(HttpStatus.CREATED.value()).body(response);
     }
 
-    @DeleteMapping("/blogs/{id}")
-    public ResponseEntity<Response> deleteBlog(@PathVariable ObjectId id) {
-        blogRepository.deleteById(id);
-        Response response = new Response(HttpStatus.OK,"Deleted Blog successfully.");
-        return ResponseEntity.status(HttpStatus.OK).body(response);
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Response> deleteBlog(@PathVariable ObjectId id, Authentication authentication) {
+        System.out.println("CALLING  DEL REQ");
+        Blog blog = blogRepository.findById(id).get();
+        if (blog.getUser().getUsername().equalsIgnoreCase(authentication.getName())) {
+            blogRepository.deleteById(id);
+            Response response = new Response(HttpStatus.OK,"Deleted Blog successfully.");
+            return ResponseEntity.status(HttpStatus.OK).body(response);
+        } else {
+            Response response = new Response(HttpStatus.UNAUTHORIZED,"User not Authrorized for this Operation");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(response);
+        }
     }
 
-    @GetMapping
+    @GetMapping("/test")
     public ResponseEntity<Response> getTestResponse() {
         Response response = new Response(HttpStatus.OK,"Test Message.");
         return ResponseEntity.status(HttpStatus.OK).body(response);
